Use usize for timeline length

+++ src/config.rs
@@ -8,7 +8,7 @@ pub struct Config {
     pub twturl: String,
     pub pre_tweet_hook: String,
     pub post_tweet_hook: String,
-    pub limit_timeline: i32,
+    pub limit_timeline: usize,
     pub use_abs_time: bool,
     pub following: HashMap<String, String>,
 }
@@ -35,7 +35,7 @@ impl Config {
             twturl: twtxt_config["twturl"].to_owned(),
             pre_tweet_hook: pre_tweet_hook,
             post_tweet_hook: post_tweet_hook,
-            limit_timeline: twtxt_config["limit_timeline"].parse::<i32>().unwrap(),
+            limit_timeline: twtxt_config["limit_timeline"].parse::<usize>().unwrap(),
             use_abs_time: twtxt_config["use_abs_time"].parse::<bool>().unwrap(),
             following: following,
         }
